requirements.txt
absl-py==1.4.0
asgiref==3.6.0
astunparse==1.6.3
async-timeout==4.0.2
cachetools==5.3.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.1.0
cloudpickle==2.2.1
Cython==0.29.34
defusedxml==0.7.1
diff-match-patch==20200713
Django==2.2.28
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-extensions==2.2.9
django-filter==23.1
django-import-export==3.1.0
django-js-asset==2.0.0
django-redis==5.2.0
django-simpleui==2023.3.1
django-sslserver2==0.1.4
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
flatbuffers==23.3.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.53.0
gym==0.26.2
gym-notices==0.0.8
h5py==3.8.0
idna==3.4
importlib-metadata==6.3.0
jax==0.4.8
Jinja2==3.1.2
keras==2.12.0
libclang==16.0.0
Markdown==3.4.3
MarkupPy==1.14
MarkupSafe==2.1.2
ml-dtypes==0.1.0
nes-py==8.2.1
numpy==1.23.5
oauthlib==3.2.2
odfpy==1.4.1
opencv-python==4.5.5.62
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
pandas==1.2.0
Pillow==9.5.0
prettytable==3.6.0
protobuf==4.22.3
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydicom==1.4.2
pyecharts==2.0.2
pyglet==1.5.21
PyMySQL==1.0.2
pypng==0.20220715.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
qrcode==7.4.2
redis==4.5.4
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
simplejson==3.18.4
six==1.16.0
smartchart==6.3.3.2
smartdb==0.6
sqlparse==0.4.3
tablib==3.4.0
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.14.1
xlrd==1.2.0
xlwt==1.3.0
zipp==3.15.0
